Interbank TextField
Se agrega la forma de un textField de Interbank para mantener en entero las monedas

+++ Flor Shop/Domain/Entities/Car.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct Car: Identifiable {
     let id: UUID
-    let total: Double
+    let total: Int
 }
